feat: :construction: The filtering process has been rebuild and it is almost complete. More test are needed more tests but the general proposal has already been built
The filtering process has been rebuilt. It was necessary to create a new table to centralized the transactions and make this table the source of the queries for the transactions list and filtering. Is necessary to make more test but the general proposal has been built and the endpoint is working fine.