refactor: :passport_control: The token-based authorization using access tokens and refresh tokens has been changed.
A new endpoint has now been created to generate access tokens using the refresh token. This endpoint will be used only when the client detects a 401 code because the access token has been declined and a new one is needed. Therefore, the user will use the endpoint, sending the refresh token via cookies to generate a new access token.